# [cuda] Parallelize random state initialization in ti.init()