--- a/taichi/runtime/llvm/llvm_runtime_executor.cpp
+++ b/taichi/runtime/llvm/llvm_runtime_executor.cpp
@@ -569,12 +569,12 @@ void LlvmRuntimeExecutor::materialize_runtime(MemoryPool *memory_pool,
     num_rand_states = config_->cpu_max_num_threads;
   }
 
-  TI_TRACE("Allocating {} random states (used by CUDA only)", num_rand_states);
+  TI_TRACE("Launching runtime_initialize");
 
-  runtime_jit->call<void *, void *, std::size_t, void *, int, int, void *,
+  runtime_jit->call<void *, void *, std::size_t, void *, int, void *,
                     void *, void *>(
       "runtime_initialize", *result_buffer_ptr, memory_pool, prealloc_size,
-      preallocated_device_buffer_, starting_rand_state, num_rand_states,
+      preallocated_device_buffer_, num_rand_states,
       (void *)&taichi_allocate_aligned, (void *)std::printf,
       (void *)std::vsnprintf);
 
@@ -582,6 +582,18 @@ void LlvmRuntimeExecutor::materialize_runtime(MemoryPool *memory_pool,
   llvm_runtime_ = fetch_result<void *>(taichi_result_buffer_ret_value_id,
                                        *result_buffer_ptr);
   TI_TRACE("LLVMRuntime pointer fetched");
+
+  if (config_->arch == Arch::cuda) {
+    TI_TRACE("Initializing {} random states using CUDA", num_rand_states);
+    runtime_jit->launch<void *, int>("runtime_initialize_rand_states_cuda",
+                                config_->saturating_grid_dim,
+                                config_->max_block_dim,
+                                0,
+                                llvm_runtime_, starting_rand_state);
+  } else {
+    TI_TRACE("Initializing {} random states (serially)", num_rand_states);
+    runtime_jit->call<void *, int>("runtime_initialize_rand_states_serial", llvm_runtime_, starting_rand_state);
+  }
 
   if (arch_use_host_memory(config_->arch)) {
     runtime_jit->call<void *>("runtime_get_mem_req_queue", llvm_runtime_);