Issue: #

### Brief Summary

This PR improves `ti.init(arch=ti.cuda)` time, especially in unit test environment.

| Concurrent processes * `ti.init()` count | Master  | With this PR | Speed up |
| --------------------------- | ------- | ------------ | -------- |
| 1 Process * 1                  | 556ms   | 410ms        | 35.6%    |
| 8 Processes * 1                 | 2500ms  | 1000ms       | 150%     |
| 1 Process * 10                 | 5850ms  | 4380ms       | 33.6%    |
| 8 Processes * 10               | 15200ms | 5760ms       | 164%     |

Filling rand states cost a whopping 150ms time during `ti.init(arch=ti.cuda)` because rand states are filled serially by a for loop in `runtime_initialize` by a `<<<1, 1>>>` kernel. This PR try to utilize GPU power to fill the rand states by a  `<<<saturating_grid_dim, max_block_dim>>>` kernel.